changes for code reveiw
Moved test to gracenote_tests, changed width calculation to include left and right.
