+++ src/annotation.ts
@@ -149,7 +149,10 @@ export class Annotation extends Modifier {
         annotation.setTextLine(state.text_line);
       }
     }
-    const overlap = Math.min(Math.max(width - maxGlyphWidth, 0), Math.max(width - state.left_shift * 2, 0));
+    const overlap = Math.min(
+      Math.max(width - maxGlyphWidth, 0),
+      Math.max(width - (state.left_shift + state.right_shift), 0)
+    );
     state.left_shift += overlap / 2;
     state.right_shift += overlap / 2;
     return true;
