+++ tests/gracenote_tests.ts
@@ -7,6 +7,9 @@
 
 import { TestOptions, VexFlowTests } from './vexflow_test_helpers';
 
+import { Accidental } from '../src/accidental';
+import { Annotation } from '../src/annotation';
+import { Articulation } from '../src/articulation';
 import { Beam } from '../src/beam';
 import { Dot } from '../src/dot';
 import { Factory } from '../src/factory';
@@ -19,6 +22,7 @@ const GraceNoteTests = {
     QUnit.module('Grace Notes');
     const run = VexFlowTests.runTests;
     run('Grace Note Basic', basic);
+    run('With Articulation and Annotation on Parent Note', graceNoteModifiers);
     run('Grace Note Basic with Slurs', basicSlurred);
     run('Grace Note Stem', stem);
     run('Grace Note Stem with Beams 1', stemWithBeamed, {
@@ -106,7 +110,43 @@ function basic(options: TestOptions): void {
 
   ok(true, 'GraceNoteBasic');
 }
+function graceNoteModifiers(options: TestOptions): void {
+  const f = VexFlowTests.makeFactory(options, 700, 130);
+  const stave = f.Stave({ x: 10, y: 10, width: 650 });
 
+  const gracenotes = [{ keys: ['b/4'], duration: '8', slash: false }].map(f.GraceNote.bind(f));
+
+  const notes = [
+    f
+      .StaveNote({ keys: ['c/5'], duration: '4', auto_stem: true })
+      .addModifier(f.GraceNoteGroup({ notes: gracenotes }).beamNotes(), 0)
+      .addModifier(new Accidental('#')),
+    f
+      .StaveNote({ keys: ['c/5'], duration: '8', auto_stem: true })
+      .addModifier(f.GraceNoteGroup({ notes: gracenotes }).beamNotes(), 0),
+    f
+      .StaveNote({ keys: ['c/5'], duration: '4', auto_stem: true })
+      .addModifier(f.GraceNoteGroup({ notes: gracenotes }).beamNotes(), 0),
+    f
+      .StaveNote({ keys: ['c/5'], duration: '4', auto_stem: true })
+      .addModifier(f.GraceNoteGroup({ notes: gracenotes }).beamNotes(), 0),
+  ];
+  notes[0].addModifier(new Articulation('a-').setPosition(3), 0);
+  notes[1].addModifier(new Articulation('a-').setPosition(3), 0).addModifier(new Annotation('words'));
+  notes[2].addModifier(new Articulation('a-').setPosition(3), 0);
+  notes[3].addModifier(new Articulation('a-').setPosition(3), 0);
+  notes[2].addModifier(new Articulation('a>').setPosition(3), 0);
+  notes[3].addModifier(new Articulation('a>').setPosition(3), 0);
+  notes[3].addModifier(new Articulation('a@a').setPosition(3), 0);
+
+  const voice = f.Voice().setStrict(false).addTickables(notes);
+
+  new Formatter().joinVoices([voice]).formatToStave([voice], stave);
+
+  f.draw();
+
+  ok(true, 'GraceNoteBasic');
+}
 function basicSlurred(options: TestOptions): void {
   const f = VexFlowTests.makeFactory(options, 700, 130);
   const stave = f.Stave({ x: 10, y: 10, width: 650 });
